+++ include/SoundManager.hpp
@@ -12,18 +12,25 @@
 #include <SFML/Audio.hpp>
 #include "Player.hpp"
 #include "TextLoader.hpp"
-#include <memory>
+#include "Events/PlaySoundEvent.hpp"
 
-using namespace std;
+#include <memory>
+#include <vector>
+#include <cassert>
+#include <random>
+#include <map>
+
+using std::string;
+using std::vector;
 
 class SoundManager{
 private:
 
   //Stores game objects referencing streaming music.
-  vector<sf::Music>music_objs;
+  std::unordered_map<int,vector<sf::Music*>>music_objs;
 
   //Stores game objects referencing pre-loaded sounds.
-  vector<sf::Sound>sound_objs;
+  std::unordered_map<std::string,sf::Sound>sound_objs;
 
   //event manager (used to register, deregister from events, and create them)
   shared_ptr<EventManager> eventManager;
@@ -33,7 +40,18 @@ private:
     A SoundBuffer to keep all of the game's sounds loaded into memory before
     they're played.
   */
-  sf::SoundBuffer buffer;
+  std::unordered_map<string,sf::SoundBuffer>buffers;
+
+  void loadSound(string path, string soundID);
+
+  int playingType;
+  unsigned int playingIndex;
+
+  const int COMBAT=0;
+  const int PREP=1;
+  const int WIN=2;
+
+  std::random_device rd;
 
 public:
 
@@ -47,12 +65,15 @@ public:
   void registerDelegates();
   void deregisterDelegates();
 
+  void loadSounds();
+
+
   /*
     Play the sound stored at the given index in sound_objs.
 
     @param dex the index in sound_objs where the sound is stored.
    */
-  void playSound(int dex);
+  void playSound(string soundID);
 
   /*
     Pause the music from music_objs at the given index if it is currently
@@ -61,7 +82,7 @@ public:
 
     @param dex the index in sound_objs where the sound is stored.
    */
-  void pauseSound(int dex);
+  void pauseSound(string soundID);
 
   /*
     Stop the music from music_objs at the given index if it is currently
@@ -70,14 +91,14 @@ public:
 
     @param dex the index in sound_objs where the sound is stored.
    */
-  void stopSound(int dex);
+  void stopSound(string soundID);
 
   /*
     Play the music stored at the given index in music_objs.
 
     @param dex the index in music_objs where the music is stored.
    */
-  void playMusic(int dex);
+  void playMusic();
 
   /*
     Pause the music from music_objs at the given index if it is currently
@@ -86,7 +107,7 @@ public:
 
     @param dex the index in music_objs where the music is stored.
    */
-  void pauseMusic(int dex);
+  void pauseMusic();
 
   /*
     Stop the music from music_objs at the given index if it is currently
@@ -95,7 +116,20 @@ public:
 
     @param dex the index in music_objs where the music is stored.
    */
-  void stopMusic(int dex);
+  void stopMusic();
+
+  void startSongOfType(int type);
+
+  void handleSoundPlay(const EventInterface& event);
+
+  void handleLevelChanged(const EventInterface& event);
+
+  void handleTowerRemove(const EventInterface& event);
+
+  void handleTowerCreation(const EventInterface& event);
+
+  void handleWaveChange(const EventInterface & event);
+
 };
 
 #endif