+++ src/CompView.cpp
@@ -42,6 +42,9 @@ void CompView::moveUnits(float deltaS){
       auto now = high_resolution_clock::now();
       //the actual count in nanoseconds for the time
       auto nowInNano = duration_cast<nanoseconds>(now.time_since_epoch()).count();
+      //if the enemy was engaged with a unit we need to denengage them both
+      currentUnit->getEngagedUnit() -> setEngagedUnit(NULL);
+      currentUnit->setEngagedUnit(NULL);
 
       shared_ptr<EventInterface> actorDestroyed = make_shared<ActorDestroyedEvent>(currentUnit->getID(),currentUnit, deltaS);
       this -> eventManager -> queueEvent(actorDestroyed);
@@ -63,9 +66,17 @@ void CompView::moveUnits(float deltaS){
 
     if(gameLogic->isExit(currentUnit->getRow(),currentUnit->getCol())){
       shared_ptr<EventInterface> actorDestroyed = make_shared<ActorDestroyedEvent>(currentUnit->getID(),currentUnit, deltaS);
+      //if the enemy was engaged with a unit we need to denengage them both
+      currentUnit->getEngagedUnit() -> setEngagedUnit(NULL);
+      currentUnit->setEngagedUnit(NULL);
+
       shared_ptr<EventInterface> hitpointsLost = make_shared<LoseHitpointsEvent>(currentUnit->getHitpoints()*textLoader->getDouble("IDS_Percentage_Unit_Hitpoint_Player_Damage"),deltaS);
+
+      shared_ptr<EventInterface> playSound = make_shared<PlaySoundEvent>("",textLoader->getString("IDS_Unit_Escape_Noise"),deltaS);
+
       this -> eventManager -> queueEvent(actorDestroyed);
       this -> eventManager -> queueEvent(hitpointsLost);
+      this -> eventManager -> queueEvent(playSound);
     }else { //these don't need to happen if the unit hit the exit
 
       int r = currentUnit->getRow();